Fix flash of back on card load while not using align options
The process of measuring the back in place causes a brief display of
the card back before the card front displays. Because this information
is only needed when using the alignHeight or alignWidth options, we
can avoid this display of the card's back when not using those options.

This commit checks whether the card is using alignHeight or alignWidth,
and if so measures the card back the same way it currently does, but if
not immediately shows the card front instead.

Test plan:

Used in an internal project, we tested loading a FlipCard and no
longer saw the blink showing the card's back when first rendering it.